login/v2

+++ src/main/java/company/friendsdog/dogcommunity/entity/Pet.java
@@ -12,8 +12,8 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Builder
 public class Pet {
-    private int userNo;
-    private int petNo;
+    private Long userNo;
+    private Long petNo;
     private String petName;
     private int petAge;
     private String petKind;
