+++ src/test/java/company/friendsdog/dogcommunity/controller/PetProfileControllerTest.java
@@ -15,8 +15,7 @@ class PetProfileControllerTest {
     @Test
     @DisplayName("프로필 삭제")
     void delete() {
-        int petNo = 3;
+        Long petNo = 3L;
         mapper.delete(petNo);
-
     }
 }
